worker/requirements.txt
# Celery & Task Queue
celery[redis]==5.3.4
redis[hiredis]==5.0.1
flower==2.0.1

# Database
asyncpg==0.29.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.25

# Vector Database
qdrant-client==1.7.3

# Document Processing
PyPDF2==3.0.1
pypdf==4.0.1
python-docx==1.1.0
openpyxl==3.1.2
python-pptx==0.6.23
Pillow==10.2.0

# OCR
pytesseract==0.3.10
pdf2image==1.17.0

# Text Processing
sentence-transformers==2.3.1
torch==2.1.2
transformers==4.37.2

# Utilities
python-dotenv==1.0.0
tenacity==8.2.3
watchdog==4.0.0

# HTTP Client for microservices
httpx==0.26.0
aiohttp==3.9.1
